+++ docs/content/en/changelog.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `toBeFuture()` expectation
 - `toBeLastYear()` expectation
 - `toBeNextYear()` expectation
+- `toBeNextMonth()` expectation
 - `toBePast()` expectation
 - `toBeSameDayAs()` expectation
 - `toBeSameHourAs()` expectation